Personal tweak to acme's exec.
There's a "hard" flag that only invokes the shell if there are some heuristics
that make it look like the execution is non-trivial.  This is fine but it is
tries to load a .dis directly, so commands that invoke shell scripts fail
silently if they don't look "hard".  This change just always invokes the shell
to execute commands in acme.  (Every machine I run acme on is fast enough to
handle this fine.)