SDL driver for /dev/draw.
It is convenient if you want to run Inferno without running X, since SDL
can draw to the framebuffer.  (There are a lot of options for this.)

This is from years ago, but it was in its own repo due to the somewhat
inconvenient way Mercurial handles branches; now that Inferno lives in
git repos, I'm merging it in.